ENH: improve element structure

+++ src/analysis/assembler/operators/heattransfer.stiffness.h
@@ -18,7 +18,7 @@ struct HeatTransferStiffnessBase: public ActionOperator {
 		isconst = false;
 	}
 
-	void move(size_t n)
+	void move(int n)
 	{
 		stiffness += n;
 	}
